v5 Update part 2
The moment you've been waiting for!
* Fixed a bug where permissions were - for some reason - checking wrongly.
* Added a 'debug logs are spammy' message upon starting the bot.
* Changed the description of 'softclear' by one word.
* It's sad to say goodbye, however 'web' will be moving to its own repo 'TheCodingBot-Extras', which will contain that sweet-sweet dashboard soon! :)
Fixed a bug where the bot would response with the prefix when there was a everyone ping.
Fixed a bug where hidden commands would still show in help. Maybe not do that please.
Removed 'lang' folder. We'll be adding multi-language support once v5 is ready for release. :)
Removed lingering 'category' part in '8ball' command.